* develop: (432 commits)
  Making ->action() work in block add/edit
  Removing missing element reference
  buffing out interface
  Fix orphaning records on file delete
  Use previously declared date helper
  Bugfix
  Fixing issue with adding control to empty layout set
  PSR
  Add specific option to control CSS source map
  PSR
  Add source maps when generating uncompressed CSS from LESS
  Loader::helper to Core::make, Loader::db to Database::connection
  PSR
  Avoid static abstract function: Permission\Access\Entity::getAccessEntitiesForUser
  Loader::db to Database::connection
  Loader::helper to Core::make, Loader::db to Database::connection, Loader::element to View::element
  PSR
  Avoid static abstract function: Gathering\Item\Item::getListByItem
  PSR
  Avoid static abstract function: Area\Layout\Column::getByID()
  ...

+++ web/concrete/authentication/concrete/controller.php
@@ -21,10 +21,13 @@ class Controller extends AuthenticationTypeController
 
     public function deauthenticate(User $u)
     {
-        list($uID, $authType, $hash) = explode(':', $_COOKIE['ccmAuthUserHash']);
-        if ($authType == 'concrete') {
-            $db = Loader::db();
-            $db->execute('DELETE FROM authTypeConcreteCookieMap WHERE uID=? AND token=?', array($uID, $hash));
+        $cookie = array_get($_COOKIE, 'ccmAuthUserHash', '');
+        if ($cookie) {
+            list($uID, $authType, $hash) = explode(':', $cookie);
+            if ($authType == 'concrete') {
+                $db = Loader::db();
+                $db->execute('DELETE FROM authTypeConcreteCookieMap WHERE uID=? AND token=?', array($uID, $hash));
+            }
         }
     }
 
@@ -79,7 +82,9 @@ class Controller extends AuthenticationTypeController
     private function genString($a = 16)
     {
         if (function_exists('mcrypt_create_iv')) {
-            return bin2hex(mcrypt_create_iv($a, MCRYPT_DEV_URANDOM));
+            // Use /dev/urandom if available, otherwise fall back to PHP's rand.
+            // http://php.net/manual/en/function.mcrypt-create-iv.php#117047
+            return bin2hex(mcrypt_create_iv($a, MCRYPT_DEV_URANDOM|MCRYPT_RAND));
         } elseif (function_exists('openssl_random_pseudo_bytes')) {
             return bin2hex(openssl_random_pseudo_bytes($a));
         }
